rename page names
