+++ new-dti-website/components/course/DDProjects.tsx
