+++ new-dti-website/components/course/TestimonialSlider.tsx
