+++ new-dti-website/components/course/Timeline.tsx
